fix(LibCarla): resolve compile-time warnings in ROS2 integration
- Add missing return path in GetOrCreateSensor (fix -Wreturn-type)
- Add virtual destructors to BasePublisher and BaseSubscriber
- Replace C-style casts with reinterpret_cast/static_cast
- Fix integer narrowing with explicit static_cast for size_t/uint64_t
- Mark unused parameters to silence -Wunused-parameter
- Remove const from by-value return types (-Wignored-qualifiers)
- Replace C++17 structured bindings for C++14 compatibility
- Reorder constructor initializers to match member declaration order

Made-with: Cursor

+++ LibCarla/source/carla/ros2/publishers/CarlaPointCloudPublisher.cpp
@@ -20,8 +20,8 @@ bool CarlaPointCloudPublisher::WritePointCloud(int32_t seconds, uint32_t nanosec
   _impl->GetMessage()->height(height);
   _impl->GetMessage()->is_bigendian(false);
   _impl->GetMessage()->fields(fields);
-  _impl->GetMessage()->point_step(point_size);
-  _impl->GetMessage()->row_step(width * point_size);
+  _impl->GetMessage()->point_step(static_cast<uint32_t>(point_size));
+  _impl->GetMessage()->row_step(static_cast<uint32_t>(width * point_size));
   _impl->GetMessage()->is_dense(false); // True if there are not invalid points
   _impl->GetMessage()->data(std::move(data));
